fix: reach a fully clippy-clean workspace under --all-targets -D warnings
Mechanical lint sweep across forgotten-host and forgotten-engine-cli:
- Remove redundant u64::from(character.id)/u16::from(vitals.health) casts and
  unnecessary vials.health.min(u16::MAX) that clippy proves are no-ops.
- Drop dead re-exports (handle_native_depot_open, native_classic_weight_description,
  deliver/resolve trade internals, MAX_ITEM_STACK_COUNT, PersistenceError) now that
  modules import directly; gate test-only re-exports (native_diagnostic_record,
  static-creature *_and_refresh) and staged public API (handle_native_depot_open,
  probe_request, three world_combat primitives) with #[allow(dead_code)] + tracking
  comments matching the NativeRenderPreparation* convention.
- Group spawn_native_static_defeat_corpse's 8 positional args into a
  NativeDefeatCorpseRequest struct (plan rule: params struct, no suppression) and
  replace a 3-tuple Option in the session consumable path with a ConsumableSource
  struct; fix identity map, needless borrows, and &PathBuf->&Path in the CLI.
- Make OperatorBridgeConfig/BridgeRuntime pub(crate)-consistent for apply_operator_request.

cargo clippy --workspace --all-targets -- -D warnings is now clean (was ~50 errors
under this toolchain); full workspace 427 tests green.

+++ crates/forgotten-host/src/bank.rs
@@ -185,8 +185,8 @@ pub(crate) fn handle_native_bank_keyword(
     }
     let mut staged_containers = containers.clone();
     for chunk in &chunks {
-        let item = ItemInstance::new(NATIVE_CURRENCY_COINS[0].0, u16::from(*chunk))
-            .map_err(HostError::Core)?;
+        let item =
+            ItemInstance::new(NATIVE_CURRENCY_COINS[0].0, *chunk).map_err(HostError::Core)?;
         let container_ids: Vec<u8> = staged_containers.iter().map(|(id, _)| id).collect();
         let mut placed = false;
         for container_id in container_ids {
